Create Working Merg
So I figured out what parts of the code it doesn't like.  This is the original Working file where I brought over the parts from NotSoMuch, a part at a time and tested to see what parts it didn't like.  If the code has //==================== above and below it, that is the code from NotSoMuch.  The following lines are the problems:
265-267 - This is the code setting up the communication with the servo boards over I2C.